+++ src/hictk/cli/cli_dump.cpp
@@ -221,8 +221,8 @@ void Cli::transform_args_dump_subcommand() {
   auto& c = std::get<DumpConfig>(_config);
 
   // in spdlog, high numbers correspond to low log levels
-  assert(c.verbosity > 0 && c.verbosity < 5);
-  c.verbosity = static_cast<std::uint8_t>(spdlog::level::critical) - c.verbosity;
+  assert(c.verbosity > 0 && c.verbosity < 5);  // NOLINTNEXTLINE(*-narrowing-conversions)
+  c.verbosity = static_cast<std::int16_t>(spdlog::level::critical) - c.verbosity;
 
   c.format = infer_input_format(c.uri);
   if (c.format == "hic" && c.resolution == 0 && c.table == "chroms") {
