+++ src/hictk/cli/cli_rename_chromosomes.cpp
@@ -99,8 +99,8 @@ void Cli::transform_args_rename_chromosomes_subcommand() {
   auto& c = std::get<RenameChromosomesConfig>(_config);
 
   // in spdlog, high numbers correspond to low log levels
-  assert(c.verbosity > 0 && c.verbosity < 5);
-  c.verbosity = static_cast<std::uint8_t>(spdlog::level::critical) - c.verbosity;
+  assert(c.verbosity > 0 && c.verbosity < 5);  // NOLINTNEXTLINE(*-narrowing-conversions)
+  c.verbosity = static_cast<std::int16_t>(spdlog::level::critical) - c.verbosity;
 }
 
 }  // namespace hictk::tools
